requirements.txt
allure-pytest==2.6.5
allure-python-commons==2.6.5
atomicwrites==1.3.0
attrs==19.1.0
backports.functools-lru-cache==1.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
configparser==3.7.4
crayons==0.2.0
future==0.17.1
idna==2.8
more-itertools==7.0.0
pluggy==0.9.0
py==1.8.0
pytest==4.4.1
pytest-assume==1.2.2
pytest-forked==1.0.2
pytest-rerunfailures==7.0
pytest-xdist==1.28.0
requests==2.21.0
selene==1.0.0a13
selenium==3.141.0
six==1.12.0
urllib3==1.24.2
webdriver-manager==1.7
